+++ CHANGELOG.md
@@ -13,7 +13,9 @@ x.x.x - TODO
   * ...
 
 3.x.x - 2024xxxx
-  * ...
+  * Add Format class, pages and filters to select by book format
+  * Fix custom column filters and handle csv with several values
+  * Sort series custom column by extra field - see issue #124 by @Mossop
 
 3.4.6 - 20241208 Fix issues + update css link
   * Changes in config/default.php file: